Added mock_service.wait_for_interactions(wait_max_seconds, poll_interval)
Calling this will wait up to 'wait_max_seconds' for interactions to
happen, checking every 'poll_interval'. As soon as the interactions
happen the method returns. In case of timeout method also retruns
silently. (So if you want to verify if the interactions actually
happened, you still need to call verify).

This method is intended to replace 'sleep x' in tests. (Adding this in a first place was motivated by 'sleeps' in
Condor to let a browser make a request to CAS).

For example, in condor at the end of a test, now we can replace:
  sleep 2
with:
  customer_acceptance.wait_for_interactions wait_max_seconds: 5, poll_interval: 0.1

This way, the test will finish as soon as possible (rather than always
waiting 2 seconds). Also, because now we are polling, we can actually
wait longer (when browser in unusally slow), like 5 seconds, so that
flaky tests are even less likly.